requirements.txt
Flask
Flask-Cache
flask_sqlalchemy
gunicorn
# requires libpq-dev
psycopg2
geoalchemy2
celery
boto
# requires geos_c
shapely
requests
raven
blinker
Flask-WTF
Flask-Login
flask-bcrypt
csvkit
redis
pytz
datetime_truncate
pyshp
Flask-Mail
lxml
netcdf4